// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/natureOfControl.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type NatureOfControl = 0 | 1;

export const NatureOfControl = {
    NUMBER_0: 0 as NatureOfControl,
    NUMBER_1: 1 as NatureOfControl
};